refactor: Filter child pages by type in fetchPagesWithChildPages
In the fetchPagesWithChildPages function, the childPages array is now filtered to only include entries of type "Entry". This ensures that only valid child pages are returned. The change improves the structure and behavior of the MenuContent component.

Also, added a TODO comment in types.ts file to replace 'any' types with the correct types defined in the file.